+++ README.md
@@ -17,10 +17,10 @@ This project is a web page that allows users to perform Create, Update, and Dele
 
 # preview
 - 1
-![pic1](https://github.com/vaishalidewangan786/react-flow/assets/95758463/097d228d-a267-4dca-86ab-6b7c2352d817)
+![Preview](https://github-production-user-asset-6210df.s3.amazonaws.com/95758463/324102577-097d228d-a267-4dca-86ab-6b7c2352d817.png?X-Amz-Algorithm=AWS4-HMAC-SHA256&X-Amz-Credential=AKIAVCODYLSA53PQK4ZA%2F20240701%2Fus-east-1%2Fs3%2Faws4_request&X-Amz-Date=20240701T174827Z&X-Amz-Expires=300&X-Amz-Signature=248ccb718abf3f682c5b5744cd2ef27d04ab63c8048419ef0f53fa66b3bd36a2&X-Amz-SignedHeaders=host&actor_id=91745441&key_id=0&repo_id=789100809)
 
 
-https://github.com/vaishalidewangan786/react-flow/assets/95758463/cc2d6971-cc0f-4b10-8433-4002dfb0e7b4
+![Demo](https://github-production-user-asset-6210df.s3.amazonaws.com/95758463/324103048-cc2d6971-cc0f-4b10-8433-4002dfb0e7b4.mp4?X-Amz-Algorithm=AWS4-HMAC-SHA256&X-Amz-Credential=AKIAVCODYLSA53PQK4ZA%2F20240701%2Fus-east-1%2Fs3%2Faws4_request&X-Amz-Date=20240701T174919Z&X-Amz-Expires=300&X-Amz-Signature=9ca224c1c4cae02f3692f645cd18f7bbdfa8d5d3d3c123ff1d2c5b511cd2efac&X-Amz-SignedHeaders=host&actor_id=0&key_id=0&repo_id=789100809)
